refactor: migrate fieldSettings:FormItem
